october 21 WIP moving onto fetch request to back end
