+++ app/javascript/react/components/RestaurantTile.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const RestaurantTile = (props) => {
+ debugger
+  return (
+    <div>
+      <h1>
+        HI from tile
+      </h1>
+    </div>
+  )
+}
+
+export default RestaurantTile
