+++ app/javascript/react/constants/restaurantsdata.json
@@ -1,3 +1,5 @@
+
+
 [
     {
         "id": 500678894,
